% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_gift_df.R
\name{append_gift_id_to_df}
\alias{append_gift_id_to_df}
\title{Append tag to edge dataframe}
\usage{
append_gift_id_to_df(df, gift_id)
}
\arguments{
\item{df}{Dataframe with the edges (from, to)}

\item{gift_id}{tag to append to the beginning of each node}
}
\value{
dataframe with the nodes renamed
}
\description{
Append tag to edge dataframe
}
\examples{
tibble::tibble(
  from = c("root_source", "a"),
  to = c("a", "sink")
) \%>\%
append_gift_id_to_df("tag")
}
