% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_gift_df.R
\name{build_gift_df}
\alias{build_gift_df}
\title{Build the GIFT database as a dataframe}
\usage{
build_gift_df()
}
\value{
data frame
}
\description{
Build the GIFT database as a dataframe
}
\examples{
build_gift_df()
}
