% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkm2.R
\docType{data}
\name{checkm2}
\alias{checkm2}
\title{CheckM2 statistics of ten MAGs}
\format{
\subsection{\code{checkm2}}{

A sample quality report from CheckM2
}
}
\source{
\url{https://github.com/anttonalberdi/distillR/data/checkm2.rda}
}
\usage{
checkm2
}
\description{
CheckM2 statistics of ten MAGs
}
\keyword{datasets}
