% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_distill_statistics.R
\name{compute_distill_statistics}
\alias{compute_distill_statistics}
\title{Compute the statistics from the distilled genomes}
\usage{
compute_distill_statistics(identifier_vector, giftdb = distillR::GIFT_db)
}
\arguments{
\item{identifier_vector}{character vector with the KEGG, EC and peptidase
identifiers}

\item{giftdb}{Table with the GIFTs (defaults to distillR::GIFT_DB)}
}
\value{
A tibble with the statistics
}
\description{
Compute the statistics from the distilled genomes
}
\examples{
identifiers <-  c("K01580", "K00823", "K16871")
compute_distill_statistics(identifiers)
}
