% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_GIFTq.R
\name{compute_GIFTq}
\alias{compute_GIFTq}
\title{Calculates the quantitative GIFTs from a gene bundle and a gene count table}
\usage{
compute_GIFTq(definition, gene_count_table)
}
\arguments{
\item{definition}{Definition string of a given gene bundle}

\item{gene_count_table}{Table containing quantitative data of genes (e.g., expression or abundance)}
}
\value{
A list of quantitative GIFT values
}
\description{
Calculates the quantitative GIFTs from a gene bundle and a gene count table
}
\examples{
compute_GIFTq(definition,gene_count_table)
}
