% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_MCI_expression.R
\name{compute_MCI_expression}
\alias{compute_MCI_expression}
\title{Calculates the gene expression/abundance-based MCI of a metabolic pathway}
\usage{
compute_MCI_expression(definition, expression_table)
}
\arguments{
\item{definition}{Definition string of a given metabolic pathway/module}

\item{expression_table}{Table containing expression data per functon}
}
\value{
A list of expression/abundance-based MCI values
}
\description{
Calculates the gene expression/abundance-based MCI of a metabolic pathway
}
\examples{
compute_MCI_expression(definition,expression_table)
}
