% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distillq.R
\name{compute_qgift}
\alias{compute_qgift}
\title{Calculates the quantitative GIFTs from a gene bundle and a gene count table}
\usage{
compute_qgift(definition, gene_count_table)
}
\arguments{
\item{definition}{Definition string of a given gene bundle}

\item{gene_count_table}{Table containing quantitative data of genes (e.g.,
expression or abundance)}
}
\value{
A list of quantitative GIFT values
}
\description{
Calculates the quantitative GIFTs from a gene bundle and a gene count table
}
\examples{
\dontrun{compute_qGIFT(definition, gene_count_table)}
}
