% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_redundancy.R
\name{compute_redundancy}
\alias{compute_redundancy}
\title{Compute the number of annotations present in a given pathway}
\usage{
compute_redundancy(annotation_vector)
}
\arguments{
\item{annotation_vector}{A vector of annotations to compute the redundancy}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{pathway_id}: The pathway identifier
\item \code{present}: The number of annotations present in the pathway
\item \code{total}: The total number of annotations in the pathway
}
}
\description{
Compute the number of annotations present in a given pathway
}
\examples{
compute_redundancy(c("S09X", "6.1.1.4"))
}
