% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\docType{data}
\name{counts}
\alias{counts}
\title{Count table of ten MAGs and ten samples}
\format{
\subsection{\code{counts}}{

A tibble with the count data for ten MAGs and ten samples
}
}
\source{
\url{https://github.com/anttonalberdi/distillR/data/counts.rda}
}
\usage{
counts
}
\description{
Count table of ten MAGs and ten samples
}
\keyword{datasets}
