% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_step_matrix.R
\name{create_step_matrix}
\alias{create_step_matrix}
\title{Creates a matrix that specifies the hierarchical structure of the definition of a metabolic pathway/module}
\usage{
create_step_matrix(def_decomp, def_level)
}
\arguments{
\item{def_decomp}{Vector of decomposed units of a definition string of a metabolic pathway/module (produced by decompose_definition())}

\item{def_level}{Vector of the hierarchical level of decomposed units of a definition string of a metabolic pathway/module (produced by set_levels())}
}
\value{
A decomposed hierarchy matrix
}
\description{
Creates a matrix that specifies the hierarchical structure of the definition of a metabolic pathway/module
}
\examples{
create_step_matrix(definition_decomposed,definition_levels)
}
