% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_gift_df.R
\name{decouple_graph}
\alias{decouple_graph}
\title{Convert a definition into a list of nodes}
\usage{
decouple_graph(definition)
}
\arguments{
\item{definition}{A pathway definition in string format}
}
\value{
A list with all the subgraphs
}
\description{
Convert a definition into a list of nodes
}
\examples{
decouple_graph("a (b c)")
}
