% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_gift_df.R
\name{definition_to_edge_df}
\alias{definition_to_edge_df}
\title{Pipeline to go from definition to edge list, with nodes properly renamed}
\usage{
definition_to_edge_df(definition, gift_id)
}
\arguments{
\item{definition}{string with the pathway definition}

\item{gift_id}{the identity of the GIFT}
}
\value{
dataframe with the edges (from, to) of the pathway
}
\description{
Pipeline to go from definition to edge list, with nodes properly renamed
}
\examples{
definition_to_edge_df("a (b,c) (c+d)", "mygift")
}
