% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gift_to_graph.R
\name{dereplicate_graph}
\alias{dereplicate_graph}
\title{Make the node names unique inside a decoupled graph}
\usage{
dereplicate_graph(decoupled_graph)
}
\arguments{
\item{decoupled_graph}{A decoupled graph from \code{decouple_graph}}
}
\value{
A decoupled graph with unique node names
}
\description{
Make the node names unique inside a decoupled graph
}
\examples{
"a (b c)" \%>\%
  decouple_graph() \%>\%
  dereplicate_graph()
}
