% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_gift_df.R
\name{dereplicated_graph_to_adjacency_list}
\alias{dereplicated_graph_to_adjacency_list}
\title{Convert dereplicated graph to adjacency list}
\usage{
dereplicated_graph_to_adjacency_list(dereplicated_graph)
}
\arguments{
\item{dereplicated_graph}{A dereplicated graph from \code{dereplicate_graph}}
}
\value{
A list of data frames with the edges
}
\description{
Convert dereplicated graph to adjacency list
}
\examples{
"a (b,c) (c+d)" \%>\%
  plus_to_space() \%>\%
  distillR::decouple_graph() \%>\%
  dereplicate_graph() \%>\%
  dereplicated_graph_to_adjacency_list()
}
