% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill_compounds.R
\name{distill_compounds}
\alias{distill_compounds}
\title{Aggregates pathway-level MCIs into compound-level MCIs}
\usage{
distill_compounds(MCI_table, pathway_db)
}
\arguments{
\item{MCI_table}{Patheway-level MCI table}

\item{pathway_db}{Table containing definitions and metadata of metabolic functions (provided by GAMMA)}
}
\value{
A MCI table aggregated at the compound level
}
\description{
Aggregates pathway-level MCIs into compound-level MCIs
}
\examples{
distill_compounds(compound_table,pathway_table)
}
