% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill_definition.R
\name{distill_definition}
\alias{distill_definition}
\title{Generates the scores of each hierarchical level of a gene bundle required to
calculate gene presence-based GIFTs}
\usage{
distill_definition(definition, def_table, level, present)
}
\arguments{
\item{definition}{Definition string of a given metabolic pathway/module}

\item{def_table}{Decomposed hierarchy matrix produced by create_step_matrix.R}

\item{level}{Hierarchical level}

\item{present}{Vector of gene identifiers present in the genome}
}
\value{
A (partially) distilled definition string
}
\description{
Generates the scores of each hierarchical level of a gene bundle required to
calculate gene presence-based GIFTs
}
\examples{
\dontrun{
distill_definition(definition, def_table, level, present)
}
}
