% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill_definition_expression.R
\name{distill_definition_expression}
\alias{distill_definition_expression}
\title{Generates the scores of each hierarchical level of a metabolic pathway required to calculate gene expression/abundance-based MCIs}
\usage{
distill_definition_expression(
  sample,
  definition_expression,
  def_table,
  level,
  expression_table
)
}
\arguments{
\item{sample}{Code of the sample to be analysed}

\item{definition_expression}{Definition-expression string}

\item{def_table}{Decomposed hierarchy matrix produced by create_step_matrix.R}

\item{level}{Hierarchical level of the pathway definition}

\item{expression_table}{Table of expression values of functional units present in the genome}
}
\value{
A (partially) distilled definition string
}
\description{
Generates the scores of each hierarchical level of a metabolic pathway required to calculate gene expression/abundance-based MCIs
}
\examples{
distill_definition_expression(definition_expression, def_table, level, expression_table)
}
