% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill_expression.R
\name{distill_expression}
\alias{distill_expression}
\title{Generates a gene abundance/expression-based Metabolic Capacity Index (MCI) table from a bacterial genome annotation table and a related gene expression table}
\usage{
distill_expression(
  expression_table,
  annotation_table,
  pathway_db,
  genecol = 1,
  genomecol = 2,
  keggcol = 9,
  eccol = c(10, 19),
  pepcol = 12
)
}
\arguments{
\item{expression_table}{Table containing normalised gene expression data with gene identifiers in rows and sample identifiers in columns}

\item{annotation_table}{Table containing gene and genome identifiers, and gene annotations}

\item{pathway_db}{Table containing definitions and metadata of metabolic functions (provided by DAMMA)}

\item{genecol}{Column index (number) of the annotations table containing the gene identifiers}

\item{genomecol}{Column index (number) of annotation_table containing the genome identifiers}

\item{keggcol}{Column index(es) of annotation_table in which to search for KEGG KO annotations}

\item{eccol}{Column index(es) of annotation_table in which to search for Enzyme Commision (EC) annotations}

\item{pepcol}{Column index(es) of annotation_table in which to search for Peptidase annotations}
}
\value{
A list of pathway-expression matrices (one table per genome)
}
\description{
Generates a gene abundance/expression-based Metabolic Capacity Index (MCI) table from a bacterial genome annotation table and a related gene expression table
}
\examples{
distill_expression(expression_table,annotation_table,pathway_db,genecol,genomecol,keggcol,eccol,pepcol)
distill_expression(expression_table,annotation_table,pathway_db,genecol=1,genomecol=2,keggcol=9,eccol=c(10,19),pepcol=12)
}
