% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill_functions.R
\name{distill_functions}
\alias{distill_functions}
\title{Aggregates compound-level MCIs into function-level MCIs}
\usage{
distill_functions(compounds_table, pathway_db)
}
\arguments{
\item{pathway_db}{Table containing definitions and metadata of metabolic functions (provided by GAMMA)}

\item{MCI_table}{Compound-level MCI table}

\item{transform}{Whether to transform to 0-1 scale. Default=TRUE}
}
\value{
A MCI table aggregated at the compound level
}
\description{
Aggregates compound-level MCIs into function-level MCIs
}
\examples{
distill_functions(compound_table)
}
