% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill_total.R
\name{distill_total}
\alias{distill_total}
\title{Aggregates compound-level MCIs into function-level MCIs}
\usage{
distill_total(functions_table, pathway_db)
}
\arguments{
\item{functions_table}{Function-level MCI table}

\item{pathway_db}{Table containing definitions and metadata of metabolic functions (provided by GAMMA)}
}
\value{
A table with Biosynthesis, Degradation and Overall MCI values
}
\description{
Aggregates compound-level MCIs into function-level MCIs
}
\examples{
distill_total(functions_table,pathway_db)
}
