% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distillq.R
\name{distillq_definition}
\alias{distillq_definition}
\title{Generates the scores of each hierarchical level of a gene bundle required to calculate gene count-based GIFTs}
\usage{
distillq_definition(
  sample,
  definition_expression,
  def_table,
  level,
  gene_count_table
)
}
\arguments{
\item{sample}{Code of the sample to be processed}

\item{definition_expression}{Definition-expression string}

\item{def_table}{Decomposed hierarchy matrix produced by create_step_matrix.R}

\item{level}{Hierarchical level of the pathway definition}

\item{gene_count_table}{Table containing quantitative data of genes (e.g., expression or abundance)}
}
\value{
A (partially) distilled definition string
}
\description{
Generates the scores of each hierarchical level of a gene bundle required to calculate gene count-based GIFTs
}
\examples{
\dontrun{
  distill_definition_expression(definition_expression, def_table, level, gene_count_table)
}
}
