% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dram.R
\docType{data}
\name{dram}
\alias{dram}
\title{DRAM annotation of ten MAGs}
\format{
\subsection{\code{dram}}{

A data frame with 10 rows and 3 columns:
}
}
\source{
\url{https://github.com/anttonalberdi/distillR/data/dram.rda}
}
\usage{
dram
}
\description{
DRAM annotation of ten MAGs
}
\keyword{datasets}
