% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_first_parenthesis.R
\name{extract_first_parenthesis}
\alias{extract_first_parenthesis}
\title{Extract the deepest-leftest parenthesis}
\usage{
extract_first_parenthesis(definition)
}
\arguments{
\item{definition}{A definition of a pathway}
}
\value{
The first parenthesis found
}
\description{
Extract the deepest-leftest parenthesis
}
\examples{
extract_first_parenthesis("a (b (c d)) (e f)")
}
