% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotations.R
\docType{data}
\name{gene_annotations}
\alias{gene_annotations}
\title{Gene annotation of a mock microbial community}
\format{
\subsection{\code{gene_annotations}}{

A data frame with 27,489 rows and 29 columns, the result from DRAM. The most
important columns are
\describe{
\item{gene}{The gene identifier}
\item{genome}{The genome / MAG the gene belongs to}
\item{kegg_id}{KEGG Ortholog identifiers}
\item{kegg_hit}{KEGG description of the kegg_id}
\item{peptidase_family}{Peptidase identifier}
\item{cazy_hits}{CAZY hits that contain Enzymatic Codes}

}
}
}
\source{
\url{https://github.com/anttonalberdi/distillR/data/gene_annotations.rds}
}
\usage{
gene_annotations
}
\description{
DRAM distilled table with the annotations for each gene of every MAG
}
\keyword{datasets}
