% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expression.R
\docType{data}
\name{gene_expression}
\alias{gene_expression}
\title{Gene expression of a mock microbial community}
\format{
\subsection{\code{gene_expression}}{

A data frame with 27,489 rows and 8 columns containing the counts of each
gene in a MAG
\describe{
\item{rows}{Genes}
\item{columns}{Samples}
\item{values}{Absolute discrete read counts}
\item{rownames}{Identifier of every gene in every MAG}

}
}
}
\source{
\url{https://github.com/anttonalberdi/distillR/data/gene_expression.rds}
}
\usage{
gene_expression
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
