% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_counts.R
\docType{data}
\name{genome_counts}
\alias{genome_counts}
\title{Genome counts of a mock microbial community}
\format{
\subsection{\code{genome_counts}}{

A data frame with 10 rows and 8 columns containing the counts of each
gene in a MAG
\describe{
\item{rows}{Genes}
\item{columns}{Samples}
\item{values}{Absolute discrete read counts}
\item{rownames}{Identifier of every gene in every MAG}

}
}
}
\source{
\url{https://github.com/anttonalberdi/distillR/data/genome_counts.rds}
}
\usage{
genome_counts
}
\description{
Genome counts of a mock microbial community
}
\keyword{datasets}
