% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_quality.R
\docType{data}
\name{genome_quality}
\alias{genome_quality}
\title{Genome quality of a mock microbial community}
\format{
\subsection{\code{genome_quality}}{

A data frame with 10 rows and 3 columns:
\describe{
\item{Genome}{The genome / MAG identifier}
\item{Completeness}{The completeness in \% of the genome}
\item{Contamination}{The percentage of contamination in the MAG / Genome}

}
}
}
\source{
\url{https://github.com/anttonalberdi/distillR/data/genome_quality.rds}
}
\usage{
genome_quality
}
\description{
Genome quality of a mock microbial community
}
\keyword{datasets}
