% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bundle_cost.R
\name{get_bundle_cost}
\alias{get_bundle_cost}
\title{Compute the shortest path cost for a given set of annotations}
\usage{
get_bundle_cost(annotation_vector)
}
\arguments{
\item{annotation_vector}{A vector of annotations to compute the shortest
path between the source and sink metabolite}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{pathway_id}: The pathway identifier
\item \code{length_shortest_path}: The length of the shortest path
\item \code{cost}: The number of genes missing in the shortest path
}
}
\description{
Compute the shortest path cost for a given set of annotations
}
\examples{
get_bundle_cost(c("S09X", "6.1.1.4"))
}
