% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_db.R
\docType{data}
\name{GIFT_db}
\alias{GIFT_db}
\title{GIFT_db}
\format{
\subsection{\code{GIFT_db}}{

Data frame with 315 rows and 7 columns, representing the different
hierarchies
of pathways
\describe{
\item{Code_bundle}{Bottom-level identifier of the pathway}
\item{Code_element}{Middle-level identifier of the pathway}
\item{Code_function}{Top-level identifier of the pathway}
\item{Domain}{Domain name}
\item{Function}{Function name}
\item{Element}{Element name}
\item{Definition}{Pathway definition as a graph}

}
}
}
\source{
\url{https://github.com/anttonalberdi/distillR/data/GIFT_db.R}
}
\usage{
GIFT_db
}
\description{
Genome-Inferred Functional Traits (GIFT)
}
\keyword{datasets}
