% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtdbtk.R
\docType{data}
\name{gtdbtk}
\alias{gtdbtk}
\title{Taxonomical assignation of ten MAGs with GTDBTK}
\format{
\subsection{\code{gtdbtk}}{

Taxonomical assignation of ten MAGs with GTDBTK
}
}
\source{
\url{https://github.com/anttonalberdi/distillR/data/gtdbtk.rda}
}
\usage{
gtdbtk
}
\description{
Taxonomical assignation of ten MAGs with GTDBTK
}
\keyword{datasets}
