% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dram.R
\name{import_dram}
\alias{import_dram}
\title{Import DRAM table}
\usage{
import_dram(dram_raw_table)
}
\arguments{
\item{dram_raw_table}{Dataframe with the raw DRAM table.
Must contain at least the following columns:
\itemize{
\item genome: genome name
\item gene:
\item kegg_id: KEGG Ortholog IDs (KOs)
\item kegg_hit: KEGG description with EC codes
\item cazy_hits: CAZy family hits with EC codes
\item peptidase_family: peptidase family
}

From this table the KOs, EC and peptidase codes will be exracted for further
analyses for distillR.}
}
\value{
Dataframe with the following columns:
\itemize{
\item genome: provided genome names
\item gene: provided gene names
\item annotation_type: type of annotation extracted
}
}
\description{
Import DRAM table
}
\examples{
distillR::gene_annotations \%>\% import_dram()
}
