% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_gift_df.R
\name{plus_to_space}
\alias{plus_to_space}
\title{Transform the plus sign into a space inside a definition}
\usage{
plus_to_space(definition)
}
\arguments{
\item{definition}{A pathway definition in string format}
}
\value{
A pathway definition in string format
}
\description{
Transform the plus sign into a space inside a definition
}
\examples{
plus_to_space("a+b")
}
