% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gift_to_graph.R
\name{process_comma_subdefinition}
\alias{process_comma_subdefinition}
\title{Process a subgraph definition with a comma: fan out from source to nodes,
fan in from nodes to sink}
\usage{
process_comma_subdefinition(subgraph_definition, subgraph_id)
}
\arguments{
\item{subgraph_definition}{A subgraph definition}

\item{subgraph_id}{The subgraph id}
}
\value{
A data frame with the edges
}
\description{
Process a subgraph definition with a comma: fan out from source to nodes,
fan in from nodes to sink
}
\examples{
process_comma_subdefinition("b,c", "subgraph_1")
}
