% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gift_to_graph.R
\name{process_single_node_subdefinition}
\alias{process_single_node_subdefinition}
\title{Process a single node subdefinition: source, node, sink}
\usage{
process_single_node_subdefinition(subgraph_definition, subgraph_id)
}
\arguments{
\item{subgraph_definition}{Subgraph definition}

\item{subgraph_id}{Subgraph id}
}
\value{
A data frame with the edges
}
\description{
Process a single node subdefinition: source, node, sink
}
\examples{
process_single_node_subdefinition("b", "subgraph_1")
}
