% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_gift_df.R
\name{process_space_subdefinition}
\alias{process_space_subdefinition}
\title{Process a subgraph definition with a space: source, node1, ... nodeN, sink}
\usage{
process_space_subdefinition(subgraph_definition, subgraph_id)
}
\arguments{
\item{subgraph_definition}{A subgraph definition}

\item{subgraph_id}{The subgraph id}
}
\value{
A data frame with the edges
}
\description{
Process a subgraph definition with a space: source, node1, ... nodeN, sink
}
\examples{
process_space_subdefinition("b c", "subgraph_1")
}
