% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_levels.R
\name{set_levels}
\alias{set_levels}
\title{Calculates the hierarchical levels of the definition of a metabolic
pathway/module}
\usage{
set_levels(definition_decomposed)
}
\arguments{
\item{definition_decomposed}{A decomposed definition string of a given
metabolic function (produced by decompose_definition())}
}
\value{
A vector of hierarchical levels per symbol in definition_decomposed
}
\description{
Calculates the hierarchical levels of the definition of a metabolic
pathway/module
}
\examples{
\dontrun{
set_levels(definition_decomposed)
}
}
