% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to.domains.R
\name{to.domains}
\alias{to.domains}
\title{Aggregates function-level GIFTs into the domain level}
\usage{
to.domains(functions_table, GIFT_db)
}
\arguments{
\item{functions_table}{Function-level GIFT table}

\item{GIFT_db}{Table containing definitions and metadata of GIFTs (default: database provided by distillR)}
}
\value{
A table with Biosynthesis, Degradation and Overall MCI values
}
\description{
Aggregates function-level GIFTs into the domain level
}
\examples{
to.domains(functions_table, GIFT_db)
}
