% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to.elements.R
\name{to.elements}
\alias{to.elements}
\title{Merges bundle-level GIFTs into the element level}
\usage{
to.elements(GIFT_table, GIFT_db = GIFT_db)
}
\arguments{
\item{GIFT_table}{Bundle-level GIFT table}

\item{GIFT_db}{Table containing definitions and metadata of GIFTs (default: database provided by distillR)}
}
\value{
A GIFT table aggregated at the element level
}
\description{
Merges bundle-level GIFTs into the element level
}
\examples{
to.elements(GIFT_table,GIFT_db)
}
