% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_community.R
\name{to_community}
\alias{to_community}
\title{Converts genome-xpecific GIFTs into the community level}
\usage{
to_community(gift_table, abundance_table, gift_db)
}
\arguments{
\item{gift_table}{GIFT table at any hierarchical level}

\item{abundance_table}{(Relative) abundance table with samples in columns
and Genomes in rows. Required for computing sample-specific GIFT values}

\item{gift_db}{Table containing definitions and metadata of GIFTs (default:
database provided by distillR)}
}
\value{
A matirx with community-level GIFTs per sample
}
\description{
Converts genome-xpecific GIFTs into the community level
}
\examples{
to.community(functions_table, abundance_table, gift_db)
}
