% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_domains.R
\name{to_domains}
\alias{to_domains}
\title{Aggregate domain-level GIFTs}
\usage{
to_domains(distilled)
}
\arguments{
\item{distilled}{tibble result from distill}
}
\value{
tibble with aggregated domain-level GIFTs
}
\description{
Aggregate domain-level GIFTs
}
