% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_elements.R
\name{to_elements}
\alias{to_elements}
\title{Aggregate element-level GIFTs}
\usage{
to_elements(distilled)
}
\arguments{
\item{distilled}{tibble result from distill}
}
\value{
tibble with aggregated domain-level GIFTs
}
\description{
Aggregate element-level GIFTs
}
