% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_function.R
\name{to_function}
\alias{to_function}
\title{Aggregate function-level GIFTs}
\usage{
to_function(distilled)
}
\arguments{
\item{distilled}{tibble result from distill}
}
\value{
tibble with aggregated function-level GIFTs
}
\description{
Aggregate function-level GIFTs
}
\examples{
gene_annotations \%>\%
  import_dram() \%>\%
  distill() \%>\%
  to_function()
}
