% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_functions.R
\name{to_functions}
\alias{to_functions}
\title{Aggregates element-level GIFTs into the function level}
\usage{
to_functions(elements_table, gift_db)
}
\arguments{
\item{elements_table}{Element-level GIFT table}

\item{gift_db}{Table containing definitions and metadata of GIFTs (default:
database provided by distillR)}
}
\value{
A GIFT table aggregated at the function level
}
\description{
Aggregates element-level GIFTs into the function level
}
\examples{
to.functions(elements_table, gift_db)
}
