% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gift_to_graph.R
\name{trim_intermediate_sources_and_sinks_df}
\alias{trim_intermediate_sources_and_sinks_df}
\title{Remove intermediate sources and sinks}
\usage{
trim_intermediate_sources_and_sinks_df(edge_df)
}
\arguments{
\item{edge_df}{A data frame with the edges}
}
\value{
A data frame with the edges
}
\description{
Remove intermediate sources and sinks
}
\examples{
"a (b,c) (c+d)" \%>\%
 plus_to_space() \%>\%
 decouple_graph() \%>\%
 dereplicate_graph() \%>\%
 dereplicated_graph_to_adjacency_list() \%>\%
 dplyr::bind_rows() \%>\%
 trim_intermediate_sources_and_sinks_df()
}
