% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_GIFT.R
\name{compute_GIFT}
\alias{compute_GIFT}
\title{Calculates a GIFT from a gene bundle}
\usage{
compute_GIFT(definition, present)
}
\arguments{
\item{definition}{Definition string of a given gene bundle}

\item{present}{Vector of identifiers present in the genome}
}
\value{
A GIFT value
}
\description{
Calculates a GIFT from a gene bundle
}
\examples{
compute_GIFT(definition,present)
compute_GIFT("K01580 (K13524,K07250,K00823,K16871) (K00135,(K00139,K17761))",c("K01580","K00823","K16871"))
}
