% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill.R
\name{distill}
\alias{distill}
\title{Generates a gene presence-based genome-Inferred Functional Trait (GIFT) table
at the gene bundle level from a bacterial genome annotation table}
\usage{
distill(annotation_table, giftdb, genomecol = 2, annotcol = c(9, 10, 19))
}
\arguments{
\item{annotation_table}{Table containing genome identifiers and gene
annotations}

\item{giftdb}{Table containing definitions and metadata of GIFTs (default:
database provided by distillR)}

\item{genomecol}{Column index (number) of the annotation_table containing the
genome identifiers}

\item{annotcol}{Column index(es) of the annotation_table in which to search
for gene identifiers (e.g., c(3,4,5))}
}
\value{
A gene bundle-level GIFT table
}
\description{
Generates a gene presence-based genome-Inferred Functional Trait (GIFT) table
at the gene bundle level from a bacterial genome annotation table
}
\examples{
distill(
    distillR::gene_annotations,
    distillR::GIFT_db,
    genomecol = 2,
    annotcol = c(9, 10, 19)
)
}
