% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill.R
\name{distill}
\alias{distill}
\title{Generates a gene presence-based Genome-Inferred Functional Trait (GIFT) table at the gene bundle level from a bacterial genome annotation table}
\usage{
distill(
  annotation_table,
  GIFT_db,
  genomecol = 2,
  keggcol = 9,
  eccol = c(10, 19),
  pepcol = 12
)
}
\arguments{
\item{annotation_table}{Table containing Genome identifiers and gene annotations}

\item{GIFT_db}{Table containing definitions and metadata of GIFTs (default: database provided by distillR)}

\item{genomecol}{Column index (number) of the annotation_table containing the genome identifiers}

\item{keggcol}{Column index(es) of the annotation_table in which to search for KEGG KO annotations}

\item{eccol}{Column index(es) of the annotation_table in which to search for Enzyme Commision (EC) annotations}

\item{pepcol}{Column index(es) of the annotation_table in which to search for Peptidase annotations}
}
\value{
A gene bundle-level GIFT table
}
\description{
Generates a gene presence-based Genome-Inferred Functional Trait (GIFT) table at the gene bundle level from a bacterial genome annotation table
}
\examples{
distill(annotation_table,GIFT_db,genomecol,keggcol,eccol,pepcol)
distill(annotation_table,GIFT_db,genomecol=2,keggcol=9,eccol=c(10,19),pepcol=12)
}
