% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill.R
\name{distill}
\alias{distill}
\title{Distill the annotations by computing the completeness of a pathway given
present annotations}
\usage{
distill(dram_annotations)
}
\arguments{
\item{dram_annotations}{Table containing the imported annotation from DRAM}
}
\value{
A tibble with the completeness of a pathway given present annotations
}
\description{
Distill the annotations by computing the completeness of a pathway given
present annotations
}
\examples{
dram |>
  import_dram() |>
  distill()
}
