% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dram.R
\name{import_dram}
\alias{import_dram}
\title{Import DRAM table}
\usage{
import_dram(dram_raw_table)
}
\arguments{
\item{dram_raw_table}{Dataframe with the raw DRAM table.
Must contain at least the following columns:
\itemize{
\item \code{...1}: gene name (DRAM leaves this column empty)
\item \code{fasta}: MAG identifier
\item \code{ko_id}: KEGG Ortholog IDs (KOs)
\item \code{kegg_hit}: KEGG description with EC codes
\item \code{cazy_ids}: CAZy family hits with EC codes
\item \code{peptidase_family}: peptidase family identifiers
}

From this table the KOs, EC and peptidase codes will be extracted for further
analyses for distillR.}
}
\value{
Dataframe with the following columns:
\itemize{
\item genome: provided genome names
\item gene: provided gene names
\item annotation_type: type of annotation extracted
}
}
\description{
Import DRAM table
}
\examples{
dram |> import_dram()
}
