% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_levels.R
\name{set_levels}
\alias{set_levels}
\title{Calculates the hierarchical levels of the definition of a metabolic pathway/module}
\usage{
set_levels(def_decomp)
}
\arguments{
\item{def_decomp}{A decomposed definition string of a given metabolic function (produced by decompose_definition())}
}
\value{
A vector of hierarchical levels per character in string
}
\description{
Calculates the hierarchical levels of the definition of a metabolic pathway/module
}
\examples{
set_levels(definition_decomposed)
}
