% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_domains.R
\name{to_domains}
\alias{to_domains}
\title{Aggregates function-level GIFTs into the domain level}
\usage{
to_domains(functions_table, gift_db)
}
\arguments{
\item{functions_table}{Function-level GIFT table}

\item{gift_db}{Table containing definitions and metadata of GIFTs (default:
database provided by distillR)}
}
\value{
A table with Biosynthesis, Degradation and Overall MCI values
}
\description{
Aggregates function-level GIFTs into the domain level
}
\examples{
to.domains(functions_table, gift_db)
}
