% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_distill.R
\name{to_elements}
\alias{to_elements}
\title{Merges bundle-level GIFTs into the element level}
\usage{
to_elements(gift_table, gift_db)
}
\arguments{
\item{gift_table}{Bundle-level GIFT table}

\item{gift_db}{Table containing definitions and metadata of GIFTs (default:
database provided by distillR)}
}
\value{
A GIFT table aggregated at the element level
}
\description{
Merges bundle-level GIFTs into the element level
}
\examples{
\dontrun{
to.elements(gift_table, gift_db)
}
}
