% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_functions.R
\name{to_functions}
\alias{to_functions}
\title{Aggregate function-level GIFTs}
\usage{
to_functions(distilled)
}
\arguments{
\item{distilled}{tibble result from distill}
}
\value{
tibble with aggregated function-level GIFTs
}
\description{
Aggregate function-level GIFTs
}
